requirements.txt
# Core ML Frameworks
tensorflow>=2.15.0
torch>=2.1.0
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Model Conversion
tensorflowjs>=4.11.0
onnx>=1.15.0
onnxruntime>=1.16.0

# Utilities
PyYAML>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0

# Optional: GPU support (uncomment if using NVIDIA GPU)
# tensorflow-gpu>=2.15.0
# torch-cuda>=2.1.0